refactor(byteseq): move `byteseq` definition into `byteseq` package
